Added Folio 3&4
3. Reinforcement Learning
Confused the benefits and drawbacks of SARSA and Q-Learning. Will need to revise this.

4. Dimensionality reduction
Used price as a target when it wasn't explicit in the assessment. The method was also slightly long winded as PCA could have been used, which would have also provided the opportunity to evaluate models produced. Some novel data exploration used for great insight into the data.